requirements.txt
prometheus-client
pyyaml
argparse
pyjson
astropy
requests
httpimport